[tests] Move Terminator specific tests into own file
